Extracts security setting tests to config-security.test.js.
