+++ test/unit/config/config.test.js
@@ -33,35 +33,6 @@ tap.test('should handle a directly passed minimal configuration', (t) => {
 })
 
 describe('the agent configuration', function() {
-  describe('with both high_security and security_policies_token defined', function() {
-    it('blows up', function() {
-      expect(function testInitialize() {
-        Config.initialize({
-          high_security: true,
-          security_policies_token: 'fffff'
-        })
-      }).throws()
-    })
-  })
-
-  describe('with a non-boolean truthy HSM setting', () => {
-    it('should enable high security mode', () => {
-      const applyHSM = Config.prototype._applyHighSecurity
-
-      let hsmApplied = false
-      Config.prototype._applyHighSecurity = () => {
-        hsmApplied = true
-      }
-      const config = Config.initialize({
-        high_security: 'true'
-      })
-      expect(!!config.high_security).to.be.true
-      expect(hsmApplied).to.be.true
-
-      Config.prototype._applyHighSecurity = applyHSM
-    })
-  })
-
   describe('when loading from a file', () => {
     describe('serverless mode', () => {
       it('should be false when config false', () => {
@@ -1222,166 +1193,6 @@ describe('the agent configuration', function() {
         })
 
         config.onConnect({'apdex_t': 0.1})
-      })
-    })
-  })
-
-  describe('#_getMostSecure', function() {
-    var config
-
-    beforeEach(function(done) {
-      config = new Config()
-      config.security_policies_token = 'TEST-TEST-TEST-TEST'
-      done()
-    })
-
-    it('returns the new value if the current one is undefined', function() {
-      var val = config._getMostSecure('record_sql', undefined, 'off')
-      expect(val).to.equal('off')
-    })
-
-    it('returns the most strict if it does not know either value', function() {
-      var val = config._getMostSecure('record_sql', undefined, 'dunno')
-      expect(val).to.equal('off')
-    })
-
-    it('should work as a pass through for unknown config options', function() {
-      var val = config._getMostSecure('unknown.option', undefined, 'dunno')
-      expect(val).to.equal('dunno')
-    })
-  })
-
-  describe('#applyLasp', function() {
-    let config = null
-    let policies = null
-    let agent = null
-
-    beforeEach(function(done) {
-      agent = {
-        _resetErrors: sinon.spy(),
-        _resetCustomEvents: sinon.spy(),
-        _resetQueries: sinon.spy(),
-        traces: {
-          clear: sinon.spy()
-        }
-      }
-      agent.config = config = new Config()
-      config.security_policies_token = 'TEST-TEST-TEST-TEST'
-      policies = securityPolicies()
-      done()
-    })
-
-    it('returns null if LASP is not enabled', () => {
-      config.security_policies_token = ''
-
-      const res = config.applyLasp(agent, {})
-      expect(res.payload).to.be.null
-    })
-
-    it('returns fatal response if required policy is not implemented or unknown', () => {
-      policies.job_arguments = { enabled: true, required: true }
-      policies.test = { enabled: true, required: true }
-
-      const response = config.applyLasp(agent, policies)
-      expect(response.shouldShutdownRun()).to.be.true
-    })
-
-    it('takes the most secure from local', () => {
-      config.transaction_tracer.record_sql = 'off'
-      config.attributes.include_enabled = false
-      config.strip_exception_messages.enabled = true
-      config.api.custom_events_enabled = false
-      config.api.custom_attributes_enabled = false
-
-      Object.keys(policies).forEach(function enablePolicy(key) {
-        policies[key].enabled = true
-      })
-
-      const response = config.applyLasp(agent, policies)
-      const payload = response.payload
-
-      expect(config.transaction_tracer.record_sql).to.equal('off')
-      expect(agent._resetQueries.callCount).to.equal(0)
-      expect(config.attributes.include_enabled).to.equal(false)
-      expect(agent.traces.clear.callCount).to.equal(0)
-      expect(config.strip_exception_messages.enabled).to.equal(true)
-      expect(agent._resetErrors.callCount).to.equal(0)
-      expect(config.api.custom_events_enabled).to.equal(false)
-      expect(agent._resetCustomEvents.callCount).to.equal(0)
-      expect(config.api.custom_attributes_enabled).to.equal(false)
-      Object.keys(payload).forEach(function checkPolicy(key) {
-        expect(payload[key].enabled).to.be.false
-      })
-    })
-
-    it('takes the most secure from lasp', () => {
-      config.transaction_tracer.record_sql = 'obfuscated'
-      config.attributes.include_enabled = true
-      config.strip_exception_messages.enabled = false
-      config.api.custom_events_enabled = true
-      config.api.custom_attributes_enabled = true
-
-      Object.keys(policies).forEach(function enablePolicy(key) {
-        policies[key].enabled = false
-      })
-
-      const response = config.applyLasp(agent, policies)
-      const payload = response.payload
-
-      expect(config.transaction_tracer.record_sql).to.equal('off')
-      expect(agent._resetQueries.callCount).to.equal(1)
-      expect(config.attributes.include_enabled).to.equal(false)
-      expect(config.attributes.exclude).to.deep.equal(['request.parameters.*'])
-      expect(config.strip_exception_messages.enabled).to.equal(true)
-      expect(agent._resetErrors.callCount).to.equal(1)
-      expect(config.api.custom_events_enabled).to.equal(false)
-      expect(agent._resetCustomEvents.callCount).to.equal(1)
-      expect(config.api.custom_attributes_enabled).to.equal(false)
-      expect(agent.traces.clear.callCount).to.equal(1)
-      Object.keys(payload).forEach(function checkPolicy(key) {
-        expect(payload[key].enabled).to.be.false
-      })
-    })
-
-    it('allow permissive settings', () => {
-      config.transaction_tracer.record_sql = 'obfuscated'
-      config.attributes.include_enabled = true
-      config.strip_exception_messages.enabled = false
-      config.api.custom_events_enabled = true
-      config.api.custom_attributes_enabled = true
-
-      Object.keys(policies).forEach(function enablePolicy(key) {
-        policies[key].enabled = true
-      })
-
-      const response = config.applyLasp(agent, policies)
-      const payload = response.payload
-
-      expect(config.transaction_tracer.record_sql).to.equal('obfuscated')
-      expect(config.attributes.include_enabled).to.equal(true)
-      expect(config.strip_exception_messages.enabled).to.equal(false)
-      expect(config.api.custom_events_enabled).to.equal(true)
-      expect(config.api.custom_attributes_enabled).to.equal(true)
-      Object.keys(payload).forEach(function checkPolicy(key) {
-        expect(payload[key].enabled).to.be.true
-      })
-    })
-
-    it('returns fatal response if expected policy is not received', () => {
-      delete policies.record_sql
-
-      const response = config.applyLasp(agent, policies)
-      expect(response.shouldShutdownRun()).to.be.true
-    })
-
-    it('should return known policies', () => {
-      const response = config.applyLasp(agent, policies)
-      expect(response.payload).to.deep.equal({
-        record_sql: { enabled: false, required: false },
-        attributes_include: { enabled: false, required: false },
-        allow_raw_exception_messages: { enabled: false, required: false },
-        custom_events: { enabled: false, required: false },
-        custom_parameters: { enabled: false, required: false }
       })
     })
   })
